+++ src/main/java/de/alpharogroup/collections/array/ArrayExtensions.java
@@ -37,6 +37,39 @@ public final class ArrayExtensions
 {
 
 	/**
+	 * Copy the given source array to the given destination array.<br>
+	 * <br>
+	 * Note: the given destination array have to be already initialized with the size of the
+	 * source.<br>
+	 * <br>
+	 * Example: <br>
+	 * <br>
+	 * <code>
+	 * Integer[] source = {1,2,3,4,5,6,7,8,9};<br>
+	 * Integer[] destination = new Integer[source.length];<br>
+	 * destination = ArrayExtensions.arraycopyWithSystem(source, destination);<br>
+	 * </code>
+	 * 
+	 *
+	 * @param <T>
+	 *            the generic type of the objects in the array.
+	 * @param source
+	 *            the source
+	 * @param destination
+	 *            the destination
+	 * @return the t[]
+	 */
+	public static <T> T[] arraycopyWithSystem(T[] source, T[] destination)
+	{
+		if (source == null)
+		{
+			return null;
+		}
+		System.arraycopy(source, 0, destination, 0, source.length);
+		return destination;
+	}
+
+	/**
 	 * Creates a new {@link List} from the given array. <br>
 	 * <br>
 	 * Note: This wraps only the method asList from {@link Arrays#asList(Object...)}.
@@ -277,55 +310,6 @@ public final class ArrayExtensions
 		final int lastIndex = array.length - 1;
 		final int indexOfElement = Arrays.asList(array).indexOf(element);
 		return indexOfElement == lastIndex;
-	}
-
-	/**
-	 * Factory method for create new array from the given optional elements.
-	 *
-	 * @param <T>
-	 *            the generic type of the elements
-	 * @param elements
-	 *            the optional elements that will be in the returned array.
-	 * @return the new array.
-	 * @deprecated use instead the same name method from the class {@link ArrayFactory} <br>
-	 *             <br>
-	 *             Note: This method will be removed in next minor release.
-	 */
-	@SafeVarargs
-	@Deprecated
-	public static <T> T[] newArray(final T... elements)
-	{
-		return elements;
-	}
-
-	/**
-	 * Creates a new {@link Integer} array with the given range that is defined through start and
-	 * end. For instance if the start is 5 and the end is 9 the resulted array will be [5,6,7,8,9]
-	 *
-	 * @param start
-	 *            The number to start
-	 * @param end
-	 *            The number to end minus one
-	 * @return the generated {@link Integer} array
-	 * @deprecated use instead the same name method from the class {@link ArrayFactory} <br>
-	 *             <br>
-	 *             Note: This method will be removed in next minor release.
-	 */
-	@Deprecated
-	public static Integer[] newRangeArray(final int start, final int end)
-	{
-		if (end < start)
-		{
-			throw new IllegalArgumentException(
-				"Parameter end should be greater than parameter start.");
-		}
-		final int length = end - start + 1;
-		final Integer[] array = new Integer[length];
-		for (int i = start; i <= end; i++)
-		{
-			array[i - start] = i;
-		}
-		return array;
 	}
 
 	/**