+++ src/test/java/de/alpharogroup/collections/CollectionExtensionsTest.java
@@ -32,11 +32,13 @@ import static org.testng.AssertJUnit.assertTrue;
 import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
 
 import org.meanbean.test.BeanTestException;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
+import de.alpharogroup.collections.array.ArrayFactory;
 import de.alpharogroup.collections.list.ListFactory;
 
 /**
@@ -64,6 +66,94 @@ public class CollectionExtensionsTest
 		{
 			assertTrue(expected.contains(number));
 		}
+	}
+
+	/**
+	 * Test method for {@link CollectionExtensions#equalCollections(Collection, Collection)}.
+	 */
+	@Test
+	public void testEqualCollections()
+	{
+		// TODO implement unit test cases...
+		boolean actual;
+		boolean expected;
+		String[] lineOne;
+		String[] lineTwo;
+		String[] lineThree;
+		String[] lineFour;
+		String[] lineFive;
+		String[] lineSix;
+		List<String[]> one;
+		List<String[]> other;
+		lineOne = ArrayFactory.newArray("1", "23", "3");
+		lineTwo = ArrayFactory.newArray("4", "25", "9");
+		lineThree = ArrayFactory.newArray("6", "21", "8");
+		one = ListFactory.newArrayList(lineOne, lineTwo, lineThree);
+		lineFour = ArrayFactory.newArray("1", "23", "3");
+		lineFive = ArrayFactory.newArray("4", "25", "9");
+		lineSix = ArrayFactory.newArray("6", "21", "8");
+		other = ListFactory.newArrayList(lineFour, lineFive, lineSix);
+		actual = CollectionExtensions.isEqualCollection(one, other);
+		expected = true;
+		assertEquals(expected, actual);
+		//
+		one = null;
+		other = null;
+		actual = CollectionExtensions.isEqualCollection(one, other);
+		expected = true;
+		assertEquals(expected, actual);
+		//
+		one = null;
+		other = ListFactory.newArrayList(lineFour, lineFive, lineSix);
+		actual = CollectionExtensions.isEqualCollection(one, other);
+		expected = false;
+		assertEquals(expected, actual);
+		//
+		one = ListFactory.newArrayList(lineOne, lineTwo, lineThree);
+		other = null;
+		actual = CollectionExtensions.isEqualCollection(one, other);
+		expected = false;
+		assertEquals(expected, actual);
+		//
+		one = ListFactory.newArrayList(lineOne, lineTwo);
+		other = ListFactory.newArrayList(lineFour, lineFive, lineSix);
+		actual = CollectionExtensions.isEqualCollection(one, other);
+		expected = false;
+		assertEquals(expected, actual);
+
+	}
+
+	/**
+	 * Test method for {@link CollectionExtensions#hashCode(java.util.Collection)}.
+	 */
+	@Test
+	public void testHashCodeCollectionWithArrayObjects()
+	{
+		int actual;
+		int expected;
+		String[] lineOne;
+		String[] lineTwo;
+		String[] lineThree;
+		List<String[]> lines;
+		lineOne = ArrayFactory.newArray("1", "23", "3");
+		lineTwo = ArrayFactory.newArray("4", "25", "9");
+		lineThree = ArrayFactory.newArray("6", "21", "8");
+		lines = ListFactory.newArrayList(lineOne, lineTwo, lineThree);
+
+		actual = CollectionExtensions.hashCode(lines);
+		expected = -1269378556;
+		assertEquals(expected, actual);
+
+		lineOne = ArrayFactory.newArray("John", "23", "male");
+		lineTwo = ArrayFactory.newArray("Jim", "25", "male");
+		lineThree = ArrayFactory.newArray("Mary", "21", "female");
+		lines = ListFactory.newArrayList(lineOne, lineTwo, lineThree);
+
+
+		actual = CollectionExtensions.hashCode(lines);
+		expected = -1124541430;
+		assertEquals(expected, actual);
+
 	}
 
 	/**