+++ src/test/java/de/alpharogroup/collections/set/SetExtensionsTest.java
@@ -25,102 +25,21 @@
 package de.alpharogroup.collections.set;
 
 import static org.testng.Assert.assertNotNull;
-import static org.testng.AssertJUnit.assertTrue;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Comparator;
-import java.util.List;
 import java.util.Set;
-import java.util.SortedSet;
 
 import org.meanbean.test.BeanTestException;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
-
-import de.alpharogroup.collections.list.ListExtensions;
-import de.alpharogroup.comparators.SortOrder;
-import de.alpharogroup.comparators.StringLengthComparator;
 
 /**
  * The unit test class for the class {@link SetExtensionsTest}.
  */
 public class SetExtensionsTest
 {
-
-	/**
-	 * Test for method {@link SetExtensions#newHashSet(java.util.Collection, Object...)}
-	 */
-	@SuppressWarnings("deprecation")
-	@Test
-	public void testNewHashSetCollectionObjects()
-	{
-		Set<String> set = SetExtensions.newHashSet();
-		assertTrue(set.size() == 0);
-		set.add("foo");
-		assertTrue(set.size() == 1);
-		set = SetExtensions.newHashSet(ListExtensions.newArrayList("foo", "fasel"), "foo", "bar",
-			"foo");
-		assertTrue(set.size() == 3);
-		set = SetExtensions.newHashSet(ListExtensions.newArrayList(), "foo", "bar", "foo");
-		assertTrue(set.size() == 2);
-
-	}
-
-	/**
-	 * Test for method {@link SetExtensions#newHashSet(Object...)}
-	 */
-	@SuppressWarnings("deprecation")
-	@Test
-	public void testNewHashSetObjects()
-	{
-		Set<String> set = SetExtensions.newHashSet();
-		assertTrue(set.size() == 0);
-		set = SetExtensions.newHashSet("foo", "bar", "foo");
-		assertTrue(set.size() == 2);
-	}
-
-	/**
-	 * Test for method {@link SetExtensions#newTreeSet(Collection, Object...)}
-	 */
-	@SuppressWarnings("deprecation")
-	@Test
-	public void testNewTreeSetCollectionObjects()
-	{
-		SortedSet<String> set = SetExtensions.newTreeSet();
-		assertTrue(set.size() == 0);
-		set.add("foo");
-		assertTrue(set.size() == 1);
-		set = SetExtensions.newTreeSet(ListExtensions.newArrayList("foo", "fasel"), "foo", "bar",
-			"foo");
-		assertTrue(set.size() == 3);
-		set = SetExtensions.newTreeSet(ListExtensions.newArrayList(), "foo", "bar", "foo");
-		assertTrue(set.size() == 2);
-
-	}
-
-
-	/**
-	 * Test method for {@link SetExtensions#newTreeSet(Collection, Comparator, T[])}.
-	 */
-	@SuppressWarnings("deprecation")
-	@Test
-	public void testNewTreeSetCollectionOfTComparatorOfTTArray()
-	{
-		SortedSet<String> set = SetExtensions.newTreeSet();
-		assertTrue(set.size() == 0);
-		set.add("foo");
-		assertTrue(set.size() == 1);
-		final Comparator<String> comparator = StringLengthComparator.of(SortOrder.ASCENDING);
-		final List<String> list = ListExtensions.newArrayList("foo", "fasel");
-		set = SetExtensions.newTreeSet(list, comparator, "food", "barista", "fao");
-		assertTrue(set.size() == 5);
-		set = SetExtensions.newTreeSet(ListExtensions.<String> newArrayList(), comparator, "foo",
-			"bar", "foo");
-		assertTrue(set.size() == 2);
-	}
-
 
 	/**
 	 * Test for method {@link SetExtensions#toSet(Collection)}