+++ src/test/java/de/alpharogroup/comparators/DateComparatorTest.java
@@ -24,18 +24,14 @@
  */
 package de.alpharogroup.comparators;
 
-import static org.testng.AssertJUnit.assertTrue;
-
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
 
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
+import de.alpharogroup.BaseComparatorTestCase;
+import de.alpharogroup.collections.list.ListFactory;
 import de.alpharogroup.date.CreateDateExtensions;
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
@@ -44,140 +40,134 @@ import lombok.experimental.FieldDefaults;
  * The unit test class for the class {@link DateComparator}.
  */
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class DateComparatorTest
+public class DateComparatorTest extends BaseComparatorTestCase<Date>
 {
 
-	/** For use of the expected result. */
-	boolean expected;
+	final Date past = CreateDateExtensions.newDate(2009, 3, 26, 10, 37, 04);
+	final Date before = CreateDateExtensions.newDate(2010, 3, 26, 10, 37, 04);
 
-	/** For use of the result of the comparator. */
-	int actual;
-
-	/** The comparator. */
-	Comparator<Date> comparator;
+	final Date after = CreateDateExtensions.newDate(2010, 3, 27, 10, 37, 04);
+	final Date future = CreateDateExtensions.newDate(2011, 3, 27, 10, 37, 04);
 
 	/**
-	 * Sets up method will be invoked before every unit test method
-	 *
-	 * @throws Exception
-	 *             is thrown if an exception occurs
+	 * {@inheritDoc}
 	 */
-	@BeforeMethod
-	public void setUp() throws Exception
+	@Override
+	protected List<Date> newActualList()
 	{
-		comparator = new DateComparator();
+		final List<Date> actual = ListFactory.newArrayList(before, future, past, after, null);
+		return actual;
 	}
 
 	/**
-	 * Tear down method will be invoked after every unit test method
-	 *
-	 * @throws Exception
-	 *             is thrown if an exception occurs
+	 * {@inheritDoc}
 	 */
-	@AfterMethod
-	public void tearDown() throws Exception
+	@Override
+	protected Comparator<Date> newComparator()
 	{
-		comparator = null;
+		return new DateComparator();
 	}
 
 	/**
-	 * Test method for {@link DateComparator#compare(Date, Date)}.
+	 * {@inheritDoc}
+	 */
+	@Override
+	protected List<Date> newExpectedSortedList()
+	{
+		final List<Date> expected = ListFactory.newArrayList(past, before, after, future);
+		expected.add(0, null);
+		return expected;
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	protected List<Date> newExpectedUnsortedList()
+	{
+		final List<Date> expected = ListFactory.newArrayList(before, future, past, after, null);
+		return expected;
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	protected Date newO1Equal()
+	{
+		return CreateDateExtensions.newDate(2009, 3, 26, 10, 37, 04);
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	protected Date newO1GreaterThan()
+	{
+		return CreateDateExtensions.newDate(2011, 3, 27, 10, 37, 04);
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	protected Date newO1LessThan()
+	{
+		return CreateDateExtensions.newDate(2009, 3, 26, 10, 37, 04);
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	protected Date newO2GreaterThan()
+	{
+		return CreateDateExtensions.newDate(2010, 3, 26, 10, 37, 04);
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	protected Date newO2LessThan()
+	{
+		return CreateDateExtensions.newDate(2010, 3, 26, 10, 37, 04);
+	}
+
+	/**
+	 * {@inheritDoc}
 	 */
 	@Test
 	public void testCompare()
 	{
-		final Date past = CreateDateExtensions.newDate(2009, 3, 26, 10, 37, 04);
-		final Date before = CreateDateExtensions.newDate(2010, 3, 26, 10, 37, 04);
-
-		final Date after = CreateDateExtensions.newDate(2010, 3, 27, 10, 37, 04);
-		final Date future = CreateDateExtensions.newDate(2011, 3, 27, 10, 37, 04);
-
-		final List<Date> dates = new ArrayList<>();
-
-		dates.add(before);
-		dates.add(future);
-		dates.add(past);
-		dates.add(after);
-		dates.add(null);
-		int i = 1;
-		for (final Date date : dates)
-		{
-			System.out.println(i + "." + date);
-			i++;
-		}
-		assertTrue("List object with index 0 should be 'before'-object.",
-			dates.get(0).equals(before));
-		assertTrue("List object with index 1 should be 'future'-object.",
-			dates.get(1).equals(future));
-		assertTrue("List object with index 2 should be 'past'-object.", dates.get(2).equals(past));
-		assertTrue("List object with index 3 should be 'after'-object.",
-			dates.get(3).equals(after));
-		assertTrue("List object with index 4 should be 'null'", dates.get(4) == null);
-		// Sort collection with our DateComparator...
-		Collections.sort(dates, comparator);
-		i = 1;
-		for (final Date date : dates)
-		{
-			System.out.println(i + "." + date);
-			i++;
-		}
-
-		assertTrue("List object with index 0 should be 'null'", dates.get(0) == null);
-		assertTrue("List object with index 1 should be 'past'-object.", dates.get(1).equals(past));
-		assertTrue("List object with index 2 should be 'before'-object.",
-			dates.get(2).equals(before));
-		assertTrue("List object with index 3 should be 'after'-object.",
-			dates.get(3).equals(after));
-		assertTrue("List object with index 4 should be 'future'-object.",
-			dates.get(4).equals(future));
+		super.testCompare();
 	}
 
 	/**
-	 * Test for the scenario when {@linkplain Date} objects are equal.
+	 * {@inheritDoc}
 	 */
 	@Test
 	public void testEqual()
 	{
-		final Date o1 = CreateDateExtensions.newDate(2009, 3, 26, 10, 37, 04);
-		final Date o2 = CreateDateExtensions.newDate(2009, 3, 26, 10, 37, 04);
-
-		actual = comparator.compare(o1, o2);
-
-		expected = actual == 0;
-		assertTrue(expected);
-
+		super.testEqual();
 	}
 
 	/**
-	 * Test for the scenario when an {@linkplain Date} object are greater than another
-	 * {@linkplain Date} object.
+	 * {@inheritDoc}
 	 */
 	@Test
 	public void testGreaterThan()
 	{
-		final Date o1 = CreateDateExtensions.newDate(2011, 3, 27, 10, 37, 04);
-		final Date o2 = CreateDateExtensions.newDate(2010, 3, 26, 10, 37, 04);
-
-		actual = comparator.compare(o1, o2);
-
-		expected = actual > 0;
-		assertTrue(expected);
+		super.testGreaterThan();
 	}
 
 	/**
-	 * Test for the scenario when an {@linkplain Date} object are less than another
-	 * {@linkplain Date} object.
+	 * {@inheritDoc}
 	 */
 	@Test
 	public void testLessThan()
 	{
-		final Date o1 = CreateDateExtensions.newDate(2009, 3, 26, 10, 37, 04);
-		final Date o2 = CreateDateExtensions.newDate(2010, 3, 26, 10, 37, 04);
-
-		actual = comparator.compare(o1, o2);
-
-		expected = actual < 0;
-		assertTrue(expected);
+		super.testLessThan();
 	}
 
 }